refactor(ModelSession): extract response methods into separate files
- Split ModelSession.swift into smaller, focused files
- Move respond methods to ModelSession+respond.swift
- Move streamResponse methods to ModelSession+streamResponse.swift
- Change helper method visibility from private to package for extension access
- Improves code organization and maintainability
