+++ Sources/SwiftAgent/ModelSession.swift
@@ -113,7 +113,7 @@ public final class ModelSession<Adapter: SwiftAgent.Adapter & SendableMetatype, 
 
 	// MARK: - Private Response Helpers
 
-	private func processResponse<Content>(
+	package func processResponse<Content>(
 		from prompt: Transcript.Prompt,
 		generating type: Content.Type,
 		using model: Adapter.Model,
@@ -204,7 +204,7 @@ public final class ModelSession<Adapter: SwiftAgent.Adapter & SendableMetatype, 
 		}
 	}
 
-	private func processResponseStream<Content: Generable>(
+	package func processResponseStream<Content: Generable>(
 		from prompt: Transcript.Prompt,
 		generating type: Content.Type,
 		using model: Adapter.Model,
@@ -262,7 +262,7 @@ public final class ModelSession<Adapter: SwiftAgent.Adapter & SendableMetatype, 
 	///
 	/// - Returns: An array of ``PromptContextLinkPreview`` objects containing metadata
 	///   for discovered URLs. Returns empty array if no URLs are found.
-	private func fetchLinkPreviews(from input: String) async -> [PromptContextLinkPreview] {
+	package func fetchLinkPreviews(from input: String) async -> [PromptContextLinkPreview] {
 		let urls = URLMetadataProvider.extractURLs(from: input)
 		guard !urls.isEmpty else { return [] }
 
@@ -344,597 +344,6 @@ public extension ModelSession {
 		return try await AuthorizationContext.$current.withValue(context) {
 			try await perform()
 		}
-	}
-}
-
-// MARK: - String Response Methods
-
-public extension ModelSession {
-	/// Generates a text response to a string prompt.
-	///
-	/// This is the most basic response method, taking a plain string prompt and returning
-	/// generated text content. The response is automatically added to the conversation transcript.
-	///
-	/// ## Example
-	///
-	/// ```swift
-	/// let response = try await session.respond(to: "What is the capital of France?")
-	/// print(response.content) // "The capital of France is Paris."
-	/// ```
-	///
-	/// - Parameters:
-	///   - prompt: The text prompt to send to the AI model.
-	///   - model: The model to use for generation. Defaults to the adapter's default model.
-	///   - options: Optional generation parameters (temperature, max tokens, etc.).
-	///     Uses automatic options for the model if not specified.
-	///
-	/// - Returns: An ``AgentResponse`` containing the generated text, transcript entries,
-	///   and token usage information.
-	///
-	/// - Throws: ``GenerationError`` or adapter-specific errors if generation fails.
-	@discardableResult
-	func respond(
-		to prompt: String,
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-	) async throws -> Response<String> {
-		let prompt = Transcript.Prompt(input: prompt, embeddedPrompt: prompt)
-		return try await processResponse(from: prompt, generating: String.self, using: model, options: options)
-	}
-
-	/// Generates a streaming text response to a string prompt.
-	///
-	/// This method provides real-time streaming of text generation, allowing you to
-	/// process and display content as it's being generated rather than waiting for
-	/// the complete response. Each snapshot contains the current state of the response.
-	///
-	/// ## Example
-	///
-	/// ```swift
-	/// for try await snapshot in session.streamResponse(to: "Tell me a story") {
-	///   print("Current content: \(snapshot.content)")
-	/// }
-	/// ```
-	///
-	/// - Parameters:
-	///   - prompt: The text prompt to send to the AI model.
-	///   - model: The model to use for generation. Defaults to the adapter's default model.
-	///   - options: Optional generation parameters (temperature, max tokens, etc.).
-	///     Uses automatic options for the model if not specified.
-	///
-	/// - Returns: An `AsyncThrowingStream` of ``AgentSnapshot`` objects containing
-	///   the current state of the response as it's being generated.
-	///
-	/// - Throws: ``GenerationError`` or adapter-specific errors if generation fails.
-	func streamResponse(
-		to prompt: String,
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-	) -> AsyncThrowingStream<Snapshot<String>, any Error> {
-		let prompt = Transcript.Prompt(input: prompt, embeddedPrompt: prompt)
-		return processResponseStream(from: prompt, generating: String.self, using: model, options: options)
-	}
-
-	/// Generates a text response to a structured prompt.
-	///
-	/// This method accepts a `Prompt` object built using the `@PromptBuilder` DSL,
-	/// allowing for more complex prompt structures with formatting and embedded content.
-	///
-	/// ## Example
-	///
-	/// ```swift
-	/// let prompt = Prompt {
-	///   "You are a helpful assistant."
-	///   PromptTag("user-input") { "What is the weather like?" }
-	///   "Please provide a detailed response."
-	/// }
-	/// let response = try await session.respond(to: prompt)
-	/// ```
-	///
-	/// - Parameters:
-	///   - prompt: A structured prompt created with `@PromptBuilder`.
-	///   - model: The model to use for generation. Defaults to the adapter's default model.
-	///   - options: Optional generation parameters.
-	///
-	/// - Returns: An ``AgentResponse`` containing the generated text and metadata.
-	///
-	/// - Throws: ``GenerationError`` or adapter-specific errors if generation fails.
-	@discardableResult
-	func respond(
-		to prompt: Prompt,
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-	) async throws -> Response<String> {
-		try await respond(to: prompt.formatted(), using: model, options: options)
-	}
-
-	/// Generates a streaming text response to a structured prompt.
-	///
-	/// This method provides real-time streaming of text generation from a structured prompt,
-	/// allowing you to process and display content as it's being generated. Each snapshot
-	/// contains the current state of the response.
-	///
-	/// ## Example
-	///
-	/// ```swift
-	/// let prompt = Prompt {
-	///   "You are a helpful assistant."
-	///   PromptTag("user-input") { "Tell me a story" }
-	/// }
-	/// for try await snapshot in session.streamResponse(to: prompt) {
-	///   print("Current content: \(snapshot.content)")
-	/// }
-	/// ```
-	///
-	/// - Parameters:
-	///   - prompt: A structured prompt created with `@PromptBuilder`.
-	///   - model: The model to use for generation. Defaults to the adapter's default model.
-	///   - options: Optional generation parameters.
-	///
-	/// - Returns: An `AsyncThrowingStream` of ``AgentSnapshot`` objects containing
-	///   the current state of the response as it's being generated.
-	///
-	/// - Throws: ``GenerationError`` or adapter-specific errors if generation fails.
-	func streamResponse(
-		to prompt: Prompt,
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-	) -> AsyncThrowingStream<Snapshot<String>, any Error> {
-		streamResponse(to: prompt.formatted(), using: model, options: options)
-	}
-
-	/// Generates a text response using a prompt builder closure.
-	///
-	/// This method allows you to build prompts inline using the `@PromptBuilder` DSL,
-	/// providing a convenient way to create structured prompts without explicitly
-	/// constructing a ``Prompt`` object.
-	///
-	/// ## Example
-	///
-	/// ```swift
-	/// let response = try await session.respond(using: .gpt4) {
-	///   "You are an expert in Swift programming."
-	///   "Please explain the following concept:"
-	///   PromptTag("topic") { "Protocol-Oriented Programming" }
-	/// }
-	/// ```
-	///
-	/// - Parameters:
-	///   - model: The model to use for generation. Defaults to the adapter's default model.
-	///   - options: Optional generation parameters.
-	///   - prompt: A closure that builds the prompt using `@PromptBuilder` syntax.
-	///
-	/// - Returns: An ``AgentResponse`` containing the generated text and metadata.
-	///
-	/// - Throws: ``GenerationError`` or adapter-specific errors if generation fails.
-	@discardableResult
-	func respond(
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-		@PromptBuilder prompt: () throws -> Prompt,
-	) async throws -> Response<String> {
-		try await respond(to: prompt().formatted(), using: model, options: options)
-	}
-
-	func streamResponse(
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-		@PromptBuilder prompt: @Sendable () throws -> Prompt,
-	) rethrows -> AsyncThrowingStream<Snapshot<String>, any Error> {
-		try streamResponse(to: prompt().formatted(), using: model, options: options)
-	}
-}
-
-// MARK: - Structured Response Methods
-
-public extension ModelSession {
-	/// Generates a structured response of the specified type from a string prompt.
-	///
-	/// This method enables structured output generation where the AI returns data conforming
-	/// to a specific `@Generable` type. This is useful for extracting structured data,
-	/// creating objects, or getting formatted responses.
-	///
-	/// ## Example
-	///
-	/// ```swift
-	/// @Generable
-	/// struct WeatherReport {
-	///   let temperature: Double
-	///   let condition: String
-	///   let humidity: Int
-	/// }
-	///
-	/// let response = try await session.respond(
-	///   to: "Get weather for San Francisco",
-	///   generating: WeatherReport.self
-	/// )
-	/// print(response.content.temperature) // Strongly-typed access
-	/// ```
-	///
-	/// - Parameters:
-	///   - prompt: The text prompt to send to the AI model.
-	///   - type: The `Generable` type to generate. Can often be inferred from context.
-	///   - model: The model to use for generation. Defaults to the adapter's default model.
-	///   - options: Optional generation parameters.
-	///
-	/// - Returns: An ``AgentResponse`` containing the generated structured content.
-	///
-	/// - Throws: ``GenerationError`` or adapter-specific errors if generation fails.
-	@discardableResult
-	func respond<Content>(
-		to prompt: String,
-		generating type: Content.Type = Content.self,
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-	) async throws -> Response<Content> where Content: Generable {
-		let prompt = Transcript.Prompt(input: prompt, embeddedPrompt: prompt)
-		return try await processResponse(from: prompt, generating: type, using: model, options: options)
-	}
-
-	/// Generates a streaming structured response of the specified type from a string prompt.
-	///
-	/// This method provides real-time streaming of structured output generation where the AI
-	/// returns data conforming to a specific `@Generable` type. Each snapshot contains the
-	/// current state of the structured response as it's being generated.
-	///
-	/// ## Example
-	///
-	/// ```swift
-	/// @Generable
-	/// struct WeatherReport {
-	///   let temperature: Double
-	///   let condition: String
-	///   let humidity: Int
-	/// }
-	///
-	/// for try await snapshot in session.streamResponse(
-	///   to: "Get weather for San Francisco",
-	///   generating: WeatherReport.self
-	/// ) {
-	///   print("Current content: \(snapshot.content)")
-	/// }
-	/// ```
-	///
-	/// - Parameters:
-	///   - prompt: The text prompt to send to the AI model.
-	///   - type: The `Generable` type to generate. Can often be inferred from context.
-	///   - model: The model to use for generation. Defaults to the adapter's default model.
-	///   - options: Optional generation parameters.
-	///
-	/// - Returns: An `AsyncThrowingStream` of ``AgentSnapshot`` objects containing
-	///   the current state of the structured response as it's being generated.
-	///
-	/// - Throws: ``GenerationError`` or adapter-specific errors if generation fails.
-	func streamResponse<Content>(
-		to prompt: String,
-		generating type: Content.Type = Content.self,
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-	) -> AsyncThrowingStream<Snapshot<Content>, any Error> where Content: Generable {
-		let prompt = Transcript.Prompt(input: prompt, embeddedPrompt: prompt)
-		return processResponseStream(from: prompt, generating: type, using: model, options: options)
-	}
-
-	/// Generates a structured response of the specified type from a structured prompt.
-	///
-	/// Combines the power of structured prompts with structured output generation.
-	/// Use this when you need both complex prompt formatting and typed response data.
-	///
-	/// ## Example
-	///
-	/// ```swift
-	/// let prompt = Prompt {
-	///   "Extract key information from the following text:"
-	///   PromptTag("document") { documentText }
-	///   "Format the response as structured data."
-	/// }
-	///
-	/// let response = try await session.respond(
-	///   to: prompt,
-	///   generating: DocumentSummary.self
-	/// )
-	/// ```
-	///
-	/// - Parameters:
-	///   - prompt: A structured prompt created with `@PromptBuilder`.
-	///   - type: The `Generable` type to generate.
-	///   - model: The model to use for generation. Defaults to the adapter's default model.
-	///   - options: Optional generation parameters.
-	///
-	/// - Returns: An ``AgentResponse`` containing the generated structured content.
-	///
-	/// - Throws: ``GenerationError`` or adapter-specific errors if generation fails.
-	@discardableResult
-	func respond<Content>(
-		to prompt: Prompt,
-		generating type: Content.Type = Content.self,
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-	) async throws -> Response<Content> where Content: Generable {
-		try await respond(
-			to: prompt.formatted(),
-			generating: type,
-			using: model,
-			options: options,
-		)
-	}
-
-	/// Generates a streaming structured response of the specified type from a structured prompt.
-	///
-	/// This method provides real-time streaming of structured output generation from a structured
-	/// prompt, combining the power of structured prompts with structured output generation.
-	/// Each snapshot contains the current state of the structured response as it's being generated.
-	///
-	/// ## Example
-	///
-	/// ```swift
-	/// let prompt = Prompt {
-	///   "Extract key information from the following text:"
-	///   PromptTag("document") { documentText }
-	///   "Format the response as structured data."
-	/// }
-	///
-	/// for try await snapshot in session.streamResponse(
-	///   to: prompt,
-	///   generating: DocumentSummary.self
-	/// ) {
-	///   print("Current content: \(snapshot.content)")
-	/// }
-	/// ```
-	///
-	/// - Parameters:
-	///   - prompt: A structured prompt created with `@PromptBuilder`.
-	///   - type: The `Generable` type to generate.
-	///   - model: The model to use for generation. Defaults to the adapter's default model.
-	///   - options: Optional generation parameters.
-	///
-	/// - Returns: An `AsyncThrowingStream` of ``AgentSnapshot`` objects containing
-	///   the current state of the structured response as it's being generated.
-	///
-	/// - Throws: ``GenerationError`` or adapter-specific errors if generation fails.
-	func streamResponse<Content>(
-		to prompt: Prompt,
-		generating type: Content.Type = Content.self,
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-	) -> AsyncThrowingStream<Snapshot<Content>, any Error> where Content: Generable {
-		streamResponse(to: prompt.formatted(), generating: type, using: model, options: options)
-	}
-
-	/// Generates a structured response using a prompt builder closure.
-	///
-	/// Allows you to build prompts inline while generating structured output,
-	/// combining the convenience of `@PromptBuilder` with typed responses.
-	///
-	/// ## Example
-	///
-	/// ```swift
-	/// let response = try await session.respond(generating: TaskList.self) {
-	///   "Create a task list based on the following requirements:"
-	///   PromptTag("requirements") { userRequirements }
-	///   "Include priority levels and estimated completion times."
-	/// }
-	/// ```
-	///
-	/// - Parameters:
-	///   - type: The `Generable` type to generate.
-	///   - model: The model to use for generation. Defaults to the adapter's default model.
-	///   - options: Optional generation parameters.
-	///   - prompt: A closure that builds the prompt using `@PromptBuilder` syntax.
-	///
-	/// - Returns: An ``AgentResponse`` containing the generated structured content.
-	///
-	/// - Throws: ``GenerationError`` or adapter-specific errors if generation fails.
-	@discardableResult
-	func respond<Content>(
-		generating type: Content.Type = Content.self,
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-		@PromptBuilder prompt: () throws -> Prompt,
-	) async throws -> Response<Content> where Content: Generable {
-		try await respond(
-			to: prompt().formatted(),
-			generating: type,
-			using: model,
-			options: options,
-		)
-	}
-
-	/// Generates a streaming structured response using a prompt builder closure.
-	///
-	/// This method provides real-time streaming of structured output generation from a prompt
-	/// built inline using the `@PromptBuilder` DSL, combining the convenience of `@PromptBuilder`
-	/// with typed responses. Each snapshot contains the current state of the structured response
-	/// as it's being generated.
-	///
-	/// ## Example
-	///
-	/// ```swift
-	/// for try await snapshot in session.streamResponse(generating: TaskList.self) {
-	///   "Create a task list based on the following requirements:"
-	///   PromptTag("requirements") { userRequirements }
-	///   "Include priority levels and estimated completion times."
-	/// } {
-	///   print("Current content: \(snapshot.content)")
-	/// }
-	/// ```
-	///
-	/// - Parameters:
-	///   - type: The `Generable` type to generate.
-	///   - model: The model to use for generation. Defaults to the adapter's default model.
-	///   - options: Optional generation parameters.
-	///   - prompt: A closure that builds the prompt using `@PromptBuilder` syntax. Must be `@Sendable`.
-	///
-	/// - Returns: An `AsyncThrowingStream` of ``AgentSnapshot`` objects containing
-	///   the current state of the structured response as it's being generated.
-	///
-	/// - Throws: ``GenerationError`` or adapter-specific errors if generation fails.
-	func streamResponse<Content>(
-		generating type: Content.Type = Content.self,
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-		@PromptBuilder prompt: @Sendable () throws -> Prompt,
-	) rethrows -> AsyncThrowingStream<Snapshot<Content>, any Error> where Content: Generable {
-		try streamResponse(to: prompt().formatted(), generating: type, using: model, options: options)
-	}
-}
-
-// MARK: - Context-Aware Response Methods
-
-public extension ModelSession {
-	/// Generates a text response with additional context while keeping user input separate.
-	///
-	/// The method automatically extracts URLs from the input and fetches link previews,
-	/// which are included in the context alongside the provided context items.
-	///
-	/// ## Example
-	///
-	/// ```swift
-	/// let response = try await session.respond(
-	///   to: "What are the key features of SwiftUI?",
-	///   supplying: [
-	///     .documentContext("SwiftUI is a declarative framework..."),
-	///     .searchResult("SwiftUI provides state management...")
-	///   ]
-	/// ) { input, context in
-	///   "You are a helpful assistant. Use the following context to answer questions."
-	///   PromptTag("context") {
-	///     for source in context.sources {
-	///       source
-	///     }
-	///   }
-	///   PromptTag("user-question") { input }
-	/// }
-	/// ```
-	///
-	/// - Parameters:
-	///   - input: The user's input/question as a plain string.
-	///   - contextItems: Array of context sources that implement ``PromptContextSource``.
-	///   - model: The model to use for generation. Defaults to the adapter's default model.
-	///   - options: Optional generation parameters.
-	///   - prompt: A closure that builds the final prompt by combining input and context.
-	///     Receives the input string and a `PromptContext` containing sources and link previews.
-	///
-	/// - Returns: An ``AgentResponse`` containing the generated text and metadata.
-	///   The transcript entry will maintain separation between input and context.
-	///
-	/// - Throws: ``GenerationError`` or adapter-specific errors if generation fails.
-	@discardableResult
-	func respond(
-		to input: String,
-		supplying contextItems: [Context],
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-		@PromptBuilder embeddingInto prompt: @Sendable (_ input: String, _ context: PromptContext<Context>) -> Prompt,
-	) async throws -> Response<String> {
-		let linkPreviews = await fetchLinkPreviews(from: input)
-		let context = PromptContext(sources: contextItems, linkPreviews: linkPreviews)
-
-		let prompt = Transcript.Prompt(
-			input: input,
-			context: context,
-			embeddedPrompt: prompt(input, context).formatted(),
-		)
-		return try await processResponse(from: prompt, generating: String.self, using: model, options: options)
-	}
-
-	func streamResponse(
-		to input: String,
-		supplying contextItems: [Context],
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-		@PromptBuilder embeddingInto prompt: @Sendable (_ input: String, _ context: PromptContext<Context>) -> Prompt,
-	) async -> AsyncThrowingStream<Snapshot<String>, any Error> {
-		let linkPreviews = await fetchLinkPreviews(from: input)
-		let context = PromptContext(sources: contextItems, linkPreviews: linkPreviews)
-
-		let prompt = Transcript.Prompt(
-			input: input,
-			context: context,
-			embeddedPrompt: prompt(input, context).formatted(),
-		)
-		return processResponseStream(from: prompt, generating: String.self, using: model, options: options)
-	}
-
-	/// Generates a structured response with additional context while keeping user input separate.
-	///
-	/// Combines context-aware generation with structured output, allowing you to provide
-	/// supplementary information while getting back strongly-typed data. This is particularly
-	/// useful for structured data extraction from contextual information.
-	///
-	/// Like the text variant, this method automatically handles URL extraction and link preview
-	/// generation from the input text.
-	///
-	/// ## Example
-	///
-	/// ```swift
-	/// @Generable
-	/// struct ProductSummary {
-	///   let name: String
-	///   let features: [String]
-	///   let price: Double
-	/// }
-	///
-	/// let response = try await session.respond(
-	///   to: "Summarize this product",
-	///   supplying: [.productDescription(productData)],
-	///   generating: ProductSummary.self
-	/// ) { input, context in
-	///   "Extract product information from the context below:"
-	///   PromptTag("context", items: context.sources)
-	///   "User request: \(input)"
-	/// }
-	/// ```
-	///
-	/// - Parameters:
-	///   - input: The user's input/question as a plain string.
-	///   - contextItems: Array of context sources that implement ``PromptContextSource``.
-	///   - type: The `Generable` type to generate.
-	///   - model: The model to use for generation. Defaults to the adapter's default model.
-	///   - options: Optional generation parameters.
-	///   - prompt: A closure that builds the final prompt by combining input and context.
-	///
-	/// - Returns: An ``AgentResponse`` containing the generated structured content and metadata.
-	///
-	/// - Throws: ``GenerationError`` or adapter-specific errors if generation fails.
-	@discardableResult
-	func respond<Content>(
-		to input: String,
-		supplying contextItems: [Context],
-		generating type: Content.Type = Content.self,
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-		@PromptBuilder embeddingInto prompt: @Sendable (_ prompt: String, _ context: PromptContext<Context>) -> Prompt,
-	) async throws -> Response<Content> where Content: Generable {
-		let linkPreviews = await fetchLinkPreviews(from: input)
-		let context = PromptContext(sources: contextItems, linkPreviews: linkPreviews)
-
-		let prompt = Transcript.Prompt(
-			input: input,
-			context: context,
-			embeddedPrompt: prompt(input, context).formatted(),
-		)
-		return try await processResponse(from: prompt, generating: type, using: model, options: options)
-	}
-
-	@discardableResult
-	func streamResponse<Content: Generable>(
-		to input: String,
-		supplying contextItems: [Context],
-		generating type: Content.Type = Content.self,
-		using model: Adapter.Model = .default,
-		options: Adapter.GenerationOptions? = nil,
-		@PromptBuilder embeddingInto prompt: @Sendable (_ prompt: String, _ context: PromptContext<Context>) -> Prompt,
-	) async -> AsyncThrowingStream<Snapshot<Content>, any Error> {
-		let linkPreviews = await fetchLinkPreviews(from: input)
-		let context = PromptContext(sources: contextItems, linkPreviews: linkPreviews)
-
-		let prompt = Transcript.Prompt(
-			input: input,
-			context: context,
-			embeddedPrompt: prompt(input, context).formatted(),
-		)
-		return processResponseStream(from: prompt, generating: type, using: model, options: options)
 	}
 }
